decoder: truncate sint32 to 32 bits before zigzag-decode
read_sint32 zigzag-decoded the full 64-bit varint and truncated afterward.
protoc truncates to 32 bits first, so a wire varint with bits set above 32
(e.g. 0x1_FFFFFFFD, a legal non-minimal encoding) decoded to 1 instead of
-2147483647. Truncate with u32() before zigzag. Fixes 6 conformance cases
(ValidDataScalar/Repeated/SelectsLast SINT32). Valid in-range values are
unaffected, so the protoc byte-oracle still matches.

+++ decoder.v
@@ -113,7 +113,10 @@ pub fn (mut d Decoder) read_uint64() !u64 {
 }
 
 pub fn (mut d Decoder) read_sint32() !int {
-	return int(zigzag_decode(d.read_varint()!))
+	// sint32 truncates the wire varint to 32 bits BEFORE zigzag-decoding, per
+	// protoc; bits above 32 must not survive into the result (else a value like
+	// 0x1_FFFFFFFD would decode to 1 instead of -2147483647)
+	return int(zigzag_decode(u64(u32(d.read_varint()!))))
 }
 
 pub fn (mut d Decoder) read_sint64() !i64 {
